module restructure and imports refactor
