+++ src/pages/documents/DocumentViewer/DocumentViewer.css
@@ -10,7 +10,7 @@
   /* add */
   display: flex;
   align-items: center;
-  justify-content: start;
+  justify-content: flex-start;
   column-gap: 10px;
   outline: none;
 }
