+++ src/pages/documents/DocumentViewer/InternalDocumentRenderer.js
