+++ src/pages/documents/common/functions/documentReferenceCreation.js
