+++ src/pages/documents/index.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import { Alert } from 'reactstrap'
 import DocumentHistory from './DocumentHistory'
 import DocumentForm from './DocumentForm'
-import { DocumentEnums } from './enums/document-enums'
+import { DocumentEnums } from './common/enums/document-enums'
 import * as ROUTES from '../../constants/routes'
 import { redirect } from '../../constants/redirect'
 import Page404 from '../errors/Page404'
